requirements.txt
easyocr
pyarabic
pymupdf
rank-bm25
nltk
torch
sentence-transformers
torch
numpy
fastapi
uvicorn
pydantic
google-generativeai